+++ test/pantomime/test/media.clj
@@ -16,3 +16,64 @@
   (is (= media/TEXT_HTML         (media/base-type (MediaType/parse "text/html; charset=UTF-8"))))
   (is (= media/APPLICATION_XHTML (media/base-type "application/xhtml+xml; charset=UTF-8"))))
 
+
+(deftest test-multipart-predicate
+  (are [mt] (is (media/multipart? mt))
+       "multipart/form-data"
+       "multipart/mixed"
+       "multipart/*"
+       "multipart/mixed; boundary=gc0p4Jq0M2Yt08j34c0p"
+       "MULTIPART/mixed; boundary=gc0p4Jq0M2Yt08j34c0p"
+       (media/media-type-named "multipart/mixed; boundary=gc0p4Jq0M2Yt08j34c0p")))
+
+(deftest test-application-predicate
+  (are [mt] (is (media/application? mt))
+       "application/postscript"
+       "application/octet-stream"
+       "application/vnd.ms-excel"
+       "application/vnd.oasis.opendocument.text"
+       "application/vnd.oasis.opendocument.spreadsheet"
+       "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
+       "application/msword"
+       "application/atom+xml"
+       "application/ecmascript"
+       "application/EDI-X12"
+       "application/javascript"
+       "application/pdf"
+       "application/rdf+xml"
+       "application/rss+xml"
+       "application/atom+xml"
+       "application/xhtml+xml"
+       "application/x-gzip"
+       "application/zip"
+       "application/font-woff"
+       "application/x-font-woff"
+       "application/x-tar"
+       "application/x-deb"
+       (media/media-type-named "application/json")
+       (media/media-type-named "application/vnd.custom; vendor=MegaCorp")))
+
+(deftest test-image-predicate
+  (are [mt] (is (media/image? mt))
+       "image/png"
+       "image/jpeg"
+       (media/media-type-named "image/gif")
+       (media/media-type-named "image/jpeg; vendor=MegaCorp")))
+
+(deftest test-audio-predicate
+  (are [mt] (is (media/audio? mt))
+       "audio/basic"
+       "audio/vnd.wave"
+       "audio/webm"
+       "audio/x-ms-wma"
+       "audio/x-pn-realaudio-plugin"
+       (media/media-type-named "audio/mp3")))
+
+(deftest test-video-predicate
+  (are [mt] (is (media/video? mt))
+       "video/mpeg"
+       "video/mp4"
+       "video/quicktime"
+       "video/webm"
+       "video/x-ms-wmv"
+       (media/media-type-named "video/quicktime")))